+++ elisp-refs.el
@@ -784,7 +784,6 @@ search."
          (pos (get-text-property (point) 'elisp-refs-start-pos))
          (unindent (get-text-property (point) 'elisp-refs-unindented))
          (column-offset (current-column))
-         (target-offset (+ column-offset unindent))
          (line-offset -1))
     (when (null path)
       (user-error "No match here"))
@@ -802,7 +801,8 @@ search."
     ;; on in the results buffer.
     (forward-line line-offset)
     (beginning-of-line)
-    (let ((i 0))
+    (let ((target-offset (+ column-offset unindent))
+          (i 0))
       (while (< i target-offset)
         (if (looking-at "\t")
             (cl-incf i tab-width)